+++ Source/Util/MapChip.cpp
@@ -108,10 +108,10 @@ Vector2 MapChip::GetPos(int32_t x,int32_t y) const
 
 uint8_t MapChip::GetNumOfArrayElement(int32_t x, int32_t y) const
 {
-	int32_t posY = min(y,map_.size() - 1);
+	int32_t posY = min(y,int32_t(map_.size() - 1));
 	posY = max(y,0);
 
-	int32_t posX = min(x,map_[ posY ].size() - 1);
+	int32_t posX = min(x,int32_t(map_[ posY ].size() - 1));
 	posX = max(x,0);
 
 	return map_[posY][posX];
